requirements.txt
flask
pandas
numpy
gunicorn
PyYAML
sklearn
evidently
dill